requirements.txt
slackclient
slackeventsapi
Flask